refactor: move garden config
