+++ src/apps/ScopeChangeRequest/ScopeChangeRequestApp.tsx
@@ -2,7 +2,6 @@ import { ClientApi } from '@equinor/lighthouse-portal-client';
 import { setupWorkspaceSidesheet } from '../../Core/WorkSpace/src/WorkSpaceApi/Functions/setupWorkspaceSidesheet';
 import { SidesheetWrapper } from './Components/Sidesheet/SidesheetWrapper/SidesheetWrapper';
 import { ScopeChangeRequest } from './types/scopeChangeRequest';
-import { gardenConfig } from './workspaceConfig/sGarden/gardenConfig';
 import { tableConfig } from './workspaceConfig/sTable/tableConfig';
 import {
     statusBarConfig,
@@ -10,6 +9,7 @@ import {
     idResolver,
     prefetchQueriesOptions,
     filterConfig,
+    gardenConfig,
 } from './utils/config';
 
 const sidesheetCreator = setupWorkspaceSidesheet<ScopeChangeRequest, 'change'>({
