+++ src/apps/ScopeChangeRequest/utils/config/garden/ScopeChangeGardenItem.tsx
@@ -1,7 +1,7 @@
-import { ScopeChangeRequest } from '../../types/scopeChangeRequest';
-import { Item } from '../../../../components/ParkView/Styles/item';
+import { ScopeChangeRequest } from '../../../types/scopeChangeRequest';
+import { Item } from '../../../../../components/ParkView/Styles/item';
 import styled from 'styled-components';
-import { CustomItemView } from '../../../../components/ParkView/Models/gardenOptions';
+import { CustomItemView } from '@equinor/ParkView';
 
 const ScopeChangeItem = styled(Item)`
     width: 100%;
