+++ src/apps/ScopeChangeRequest/utils/config/index.ts
@@ -3,3 +3,4 @@ export * from './dataCreatorConfig';
 export * from './dataOptions';
 export * from './prefetchQueryOptions';
 export * from './filter/filterConfig';
+export * from './garden/gardenConfig';
